fix(biometrics): allow face unlock to work from backgrounded app
also allow passcode as fallback
